start/requirements.txt
openai
python-dotenv
sounddevice
numpy
scipy
openai-whisper
pyttsx3
torch